Reset state in home page
When going through the whole flow. The redux state kept information about the previous game, which meant that when starting a new game, the previous username would still display in a room. This resets the state anytime the user goes back to the home page in order to fix that bug.